Make routes directory and move routes JS files to routes directory
